Let's describe a new behavior that we need for our Dinosaur class. I want to be able to easily get a description of the Dinosaur - a string that will contain the type of dinosaur, whether or not it likes to eat people and its length.
Let's turn that into a new example: function it_should_return_full_description(). For this first example, we'll describe what the description should look like if we set no data. Let's say that there should be a new getDescription() method that shouldReturn():

  The Unknown non-carnivorous dinosaur is 0 meters long

Our Dinosaur doesn't even have any properties on it related to what 'type' of Dinosaur it is, or whether it's carnivorous or non-carnivorous - but those are details for future Ryan to worry about.