The Wonderful Lesson of Hardcoding Return Values
Ok, remember the red, green, refactor cycle? For step two, we're technically supposed to make this test green with as little work as possible. Challenge accepted! Let's copy this string, go into Dinosaur, find the new method, and - yes, I am about to do this - return that hardcoded string! We are awesome at programming! For extra credit, let's add a return type.

Yes, I do realize how silly this is. And no, I don't do this when I'm coding for real. But, there's something... beautiful about hardcoding this value: it's a reminder to focus on what we truly need to accomplish in this method - and to not over-complicate things or add extra options we don't need yet. BDD says: if you need your code to be more flexible, you'll discover that naturally when you describe some new behavior in a new example. The design of your code 'emerges' naturally after writing examples and getting them to pass.

When we try phpspec again:
 ./vendor/bin/phpspec run
Shocking! It passes