Describing Constructor Arguments
Head back to our DinosaurSpec class: to fully describe how we want the new getDescription() method to work, we need a few more examples. Let's see: the description contains details about the 'type' or 'genus' of the dinosaur - like Tyrannosaurus or Stegosaurus - and whether or not it wants to eat you. I mean, whether or not it's carnivorous.

Right now, there is no way to set this type of info on the Dinosaur class. We need to fix that. So... how do we want to set that info? For the length, we added a setLength() method. But, I think that the type of dinosaur and whether or not it's carnivorous are so important, that they should be passed via the constructor when instantiating a Dinosaur.

Let's create an example: it_should_return_full_description_for_tyrannosaurus(). We know that phpspec handles instantiating a new Dinosaur object for us so that when we call getDescription(), it eventually calls that method on the real object.