How the Objects are Constructed Oh, by the way, what if we had two beConstructedWith() lines with different arguments? I know, this looks silly - but this can happen in some cases when you use a setup function called let() that we'll learn about later.
Anyways, what would happen here? An error? CPU over-heating? Neither! The last call always wins. The reason is the interesting part. Behind the scenes, phpspec delays instantiating the object as long as it can. In this case, it doesn't actually instantiate the Dinosaur object until we call a method on it like setLength(). At that moment, phpspec realizes it needs to instantiate the object and creates it using the arguments that were passed to the last beConstructedWith() call.