phpspec just generated the __construct() method for us. Thanks buddy! Go check it out! Two cool things here. First, in Dinosaur, yes, it did add the constructor method. As a bonus, it even put it in the right place: after the properties, but above all the other public functions. phpspe, are you trying to take my job?
Second, when phpspec re-ran all of the examples, well... they're almost all failing now: too few arguments to Dinosaur::__construct(). And... that makes sense! We just massively changed the way that our Dinosaur class is designed. And so, any existing coding using that class will probably be totally borked! This is a great example of how our tests can give us feedback. They're saying:

Ryan! Do you realize that you just broke all of the code in your app that creates new Dinosaur objects?

If we did that on accident... well... that would be a pretty good warning to get.