Implementing the Constructor
Let's get to work: the first argument should be a string $ genus and the second bool $ isCarnivorous. I'll press Alt+Enter and select 'Initialize Fields'... which is a shortcut to create those two properties and set them. Remove the TODO.

This is cool. But... I'm going to make both of these arguments optional. Why? Well, it's entirely up to you how you want your class to work. Look back at DinosaurSpec. According to this example, it looks like it should be legal to create a Dinosaur object with no information. If you do, the type should be 'unknown' and it should default to not eat you... which is kinda nice. This is 'design by spec': we're using our examples to drive how the class is built.

Default $ genus to Unknown and $ isCarnivorous to false.

Using --verbose

The getDescription() method is still wrong. But, we did just get a step closer, so let's try phpspec again:
./vendor/bin/phpspec run
Yep! The two strings don't match. By the way, see how it truncates the two strings? Sometimes that makes it hard to figure out what's going on. If you need more info, run phpspec with the --verbose option:
 ./vendor/bin/phpspec run --verbose