Back in Dinosaur, let's finish the getDescription() method. Wrap the string in sprintf() then add a few wildcards: one for the genus, one for the non- part and one for the length. Fill these in with $ this->genus, a ternary to print either nothing, or non-, and then $ this->length.
Try it again:
 ./vendor/bin/phpspec run --verbose
Ah! It still fails! Whoops - I made a mistake in the spec file - but it's obvious. I've been using phpspec for so long that I can't avoid saying 'should' in everything I type. Change should be to is: that's the language we want.

Try it one more time:
./vendor/bin/phpspec run
It passes! Next: with production ramping up, we need a factory for our dinosaurs. Let's see how we can describe that with phpspec.