You know the drill: after writing the example, run phpspec: ./vendor/bin/phpspec run
Honestly, it shouldn't even be a surprise any more when phpspec generates code for us. Choose 'yes' so it generates the missing growVelociraptor() method. And when it re-executes... failure! A BadMethodCallException(). That comes from the new function. phpspec knows this should return a Dinosaur object... but it's not sure how. But hey! It did, once again, put this method in just the right place: below the constructor but above the public functions.