Change the argument to int $ length, advertise that this will return an instance of self, and create that with $ dinosaur = new static() passing Velociraptor and true for the isCarnivorous argument. Then, $ dinosaur->setLength($ length), return $ dinosaur.
That felt good! Let's make sure the example passes:
./vendor/bin/phpspec run

Wait... it failed! Of course! The getGenus() method doesn't exist! That's super cool: instead of planning ahead and adding this method earlier, we allowed the need for this method to 'emerge' naturally. What's especially interesting is that, so far, the only place we need this method is in our example! What if we don't need this method in our actual app? Should we still create it? Yes. Well, let me say that differently. The code in our examples are meant to be real examples of how you want your class to work. If you really don't want a getGenus() method, then should write example code that doesn't use it. If you do use it, you need it