That felt good! Let's make sure the example passes: ./vendor/bin/phpspec run
The code in our examples are meant to be real examples of how you want your class to work. If you really don't want a getGenus() method, then should write example code that doesn't use it. If you do use it, you need it!

phpspec, please generate that for me. Find the new method, return $ this->genus and add the string return type. Try the tests again:
./vendor/bin/phpspec run
Yes! All green! Next: as cool as this factory method is, we need to level up with a proper, new DinosaurFactory service class.