Run: ./vendor/bin/phpspec describe App/Factory/DinosaurFactory
Describing a new Class
So, hey! We need a new class! Well, to say it better, it's time for us to describe a new class: ./vendor/bin/phpspec describe and, for the name, how about App/Factory/DinosaurFactory.
That creates one new file: DinosaurFactorySpec. Let's go check it out! Like last time, we get one super basic example for free - asserting that  should be an instance of DinosaurFactory. That's... kinda silly... but it's enough to force some code generation!