growVelociraptor() Example
Our first goal is to move the growVelociraptor() method into DinosaurFactory, but I want to follow the red, green, refactor cycle. So first, describe that functionality with a new example: function it_grows_a_large_velociraptor(). Then, call the method: $ dinosaur = $ this->growVelociraptor(5).

The Magic Behind phpspec's Subject

Eventually, after coding all of this up, we know that the  variable should be a Dinosaur object. But we also know that phpspec adds a lot of magic. Check this out: var_dump($ dinosaur). Now, run phpspec:
./vendor/bin/phpspec run
First, it notices that the growVelociraptor() method is missing. Hit enter to generate that. Ok: scroll up to check out the dumped object. Cool! The  variable is actually a Subject object! Right now, the underlying value is null because the new growVelociraptor() method doesn't return anything.

First, it notices that the growVelociraptor() method is missing. Hit enter to generate that. Ok: scroll up to check out the dumped object. Cool! The $ dinosaur variable is actually a Subject object! Right now, the underlying value is null because the new growVelociraptor() method doesn't return anything.

But more importantly, do you remember where we saw the Subject object earlier? It was in DinosaurSpec! When we call $ this->getLength(), that returns the length, but wrapped inside of a Subject object. Why do we care? Because that was the magic layer that allowed us to call ->shouldReturn().

Inside DinosaurFactorySpec, it's the same thing! growVelociraptor will eventually return a Dinosaur object, but phpspec wraps that inside a Subject object. Thanks to that, we can call real methods on the Dinosaur or matcher methods. In other words, the $ dinosaur in this class works pretty much exactly like the $ this variable in DinosaurSpec.