Implement the Code
Step 2: make this work with as little work as possible... or at least without over-engineering it. We can cheat: copy the code from the old growVelociraptor() method. I'll keep this method here just as an example. Back in DinosaurFactory, paste, change the new static to new Dinosaur, change the argument to int $ length and give this a Dinosaur return type.
Try it out:
./vendor/bin/phpspec run
Green!