Until now, our examples haven't really needed to involve other objects. For example, in DinosaurSpec, when we call setLength(), we pass a scalar argument. We haven't had a situation yet where the object that we're describing depends on another object. That's something that we need to talk a lot more about.
Pending & Skipped Examples

But first, in DinosaurFactory, we have this growVelociraptor() method. Eventually we're going to add other methods to grow other things. And just to make sure I don't forget to do that, let's create a new it_grows_a_triceratops() example. But I'm not actually ready to describe this or implement it yet... I don't know.... maybe there's a big storm coming and I need to get off the island on the last boat or something. So, just leave it blank and run phpspec:
./vendor/bin/phpspec run
Cool! It shows up as a 'todo' pending example! Once we come back after the storm, we won't forget!

One other thing you can do, which is a bit less common in apps, but still neat, is to skip tests if you're missing some sort of dependency. For example, create it_grows_a_small_velociraptor().

Let's pretend like we need an outside library that contains a class called Nanny in order to create baby velociraptors. If that class doesn't exist, we can throw a new SkippingException that says:

  Someone needs to look over dino puppies

So, no Nanny class? Skip the example. If we do have it, it will run like normal: $ this->growVelociraptor(1) and, how about, ->shouldBeAnInstanceOf(Dinosaur::class).
Since that's just a made-up class, when we run phpspec:
./vendor/bin/phpspec run
Yep! That one got skipped.