Ok, good start! Head back over to your terminal and run things: ./vendor/bin/phpspec run Enter yes to generate that class, and yes to generate the getDinosaurs() method inside of it.
Thanks to that, not only do we have the new Enclosure class, but it already has its first method!

Basic Enclosure Implementation

To get started, we probably need a $ dinosaurs property, which will hold an array of Dinosaur objects. Add an array return type to the method and return $ this->dinosaurs. Oh, and let's initialize the property to an empty array - that's exactly the behavior we're describing.

We could have just returned a hardcoded empty array... because that is the minimum code we need to get the test to pass. But as you get more comfortable with phpspec, it's ok to start skipping that step - as long as you stay focused on the behavior you need and don't allow yourself to get too fancy.

Let's make sure things are passing:
./vendor/bin/phpspec run
Perfect! Just the one, pending example.