Adding Dinosaurs to the Enclosure
Let's think a bit more about the Enclosure. We will definitely need a way to to add dinosaurs to it. Let's describe that! function it_should_be_able_to_add_dinosaurs(). And because we'll most likely be adding Dinosaurs one-by-one as they're born, I think an addDinosaur() method will be quite perfect: $ this->addDinosaur() and pass that a Dinosaur object.

Mock the Dinosaur Object?

But wait! This is the first time that we're calling a method on our object and what we need to pass to that method is... another object! Ok... so what's the big deal? Remember, in unit tests, each class is supposed to be tested in complete isolation. If you have a class that depends on a database connection, instead of passing the real database connection object, you're supposed to pass it a mock object so that it doesn't make real database queries and also so we can fake and control the return value of its methods.

So... question time: should we mock the Dinosaur object? And if so, how do we do that?

The answer is... probably no: we should not mock it. Whenever you need to pass an object to the object you're testing, you need to decide whether or not to mock it. And the correct answer depends on how difficult it is to instantiate the object and control its behavior. For example, the Dinosaur object is a simple model object, and it doesn't really do anything - it just holds data. It's easy to instantiate and, if we want its getLength() method to return 7 to help us test something, yea, that's super easy! Just set its length to 7!

The point is: the Dinosaur object is so simple, that mocking will work, but it will make your life harder! That's why I prefer to pass in the real object.

If this were a database connection object, something that sent emails or any other class that did some real work, I would mock it, and we'll talk about how to mock things in phpspec soon.

Let's copy this line so we can add two dinosaurs. And then say $ this->getDinosaurs()->shouldHaveCount(2).