Sometimes your code will throw an exception... it's just how things are. Actually, sometimes it's super important that the right exception is thrown at the exact right time. Let's see an example - and then, see how to describe that behavior in a spec class.
If you downloaded the course code, you should have a tutorial/ directory. Find the Exception/ directory inside of that and copy that whole darn thing into src/. This holds two exception classes, and the first one that we're going to look at is, the very important, NotABuffetException. You see, we've had this problem where sometimes we accidentally put a veggiesaurus inside an Enclosure with a carnivorous dinosaur. And, well, the results have been... messy.

To make sure we stop doing that, we need to throw this exception if we try to mix carnivorous and non-carnivorous dinosaurs into the same Enclosure. And this is such an important thing, we need to make sure there is a test to ensure the carnage stops.