Example for an Exception
Open up EnclousreSpec: because we want the exception to be thrown when the addDinosaur() method is called. Let's say: function it_should_not_allow_to_add_carnivorous_dinosaurs_to_non_carnivorous_enclosure().

Wow! That's a long name - but... ok! It's a great description for this example.
Here's the plan: we're going to add one dinosaur that's a veggie dinosaur and then add a another dinosaur that's carnivorous. And that should trigger the exception. Start with $ this->addDinosaur(new Dinosaur()) and make this a veggie eater by passing false as the second argument.

Now, here is the important part: before we call addDinosaur(), again, we need to tell phpspec to expect that there should be an exception. And it's probably no surprise that the language to do this is really natural: $ this->shouldThrow() NotABuffetException::class, ->during(), and then we tell phpspec exactly what method should trigger this: addDinosaur with an array of the arguments, and we only have one: new Dinosaur(), Velociraptor and true for the carnivorous argument.

That's it! Let's try it out:
./vendor/bin/phpspec run
Cool! Failure because no exception was thrown.