Time for us to get to work! In addDinosaur(), we need to determine whether or not we're allowed to add this Dinosaur. Let's call a new function: if (! $ this->canAddDinosaur()) - we'll create that method in a minute - then throw new NotABuffetException().
Now I can click back on canAddDinosaur, press Alt + Enter and click 'Add Method' to create a new private method at the bottom. Oh, and I'm just creating this as a private method for code organization: I could have written the logic right up in the addDinosaur() function. But, it is nice to have a method called canAddDinosaur() - it's really clear. I'm not making it public because, at least so far, we don't have any need to use it outside of this class. That also means that we won't write an example for this function: examples are only for public methods.

Anyways, let's add a return type and then say return count($ this->dinosaurs) === 0 - because if there are no dinosaurs in this enclosure, than we can definitely add one - or we can check if the first dinosaur, index 0, has the same 'diet' as the dinosaur being added, it should be allowed. So... hmm, maybe we'll call ->isCarnivorus().

But, wait... that method does not exist in the Dinosaur class. And actually, the real problem is that the isCarnivorous information is not available publicly in any way, except as part of the description.

This is cool! We just discovered that we need to enhance the Dinosaur class to get the new feature working. Before we do that, let's finish the canAddDinosaur() method: you should be able to add the dinosaur if the first dinosaur ->isCarnivorous() value equals $ dinosaur->isCarnivorous(). If they are compatible, this dinosaur can be added to the enclosure... without being eaten... hopefully.

We know we're not done yet, but in the spirit of 'doing as little work as possible and letting phpspec tell us what to do next', let's run it now:
./vendor/bin/phpspec run
Failure!

  Call to undefined method Dinosaur::isCarnivorous().

We could now go directly into the Dinosaur class and just... implement that! It's a super-easy method. But... don't we need to write an example first before we add the code? Maybe. So far, we've been testing a lot of simple getter and setter methods. You can test simple methods like this, but at some point a method is so simple that... in my opinion, testing them is overkill. Focus on testing what scares you.

But, for our great learning adventure, we will add some examples for this method. Why? Because it will introduce us to a really cool matcher for boolean methods. Let's check it out next.