We're missing the isCarnivorous() method on Dinosaur. It will be a really simple method, but because it will let me show you a very special match - the ObjectStateMatcher - we're going to write a couple of examples for it.
Find DinosaurSpec and add the first new example: it_should_be_herbivore_by_default(). This example is meant to show that if we create a new Dinosaur() without passing the isCarnivorous argument, it should default to an herbivore. Pay careful attention to the language I'm about to use. Behind the scenes, because we haven't controlled the constructor arguments, we know that phpspec will create a Dinosaur() without any constructor arguments. So I'm going to say $ this->shouldNotBeCarnivorous().

That is not a built-in matcher... right? I mean, that's super specific language, and PhpStorm definitely did not autocomplete that for me. Well... surprise! That is a real matcher! Say hello to the ObjectStateMatcher! It's dynamic: whenever you say shouldBeSOMETHING or shouldNotBeSOMETHING(), the ObjectStateMatcher is activated. It parses out that 'SOMETHING' part - for us the word Carnivorous - looks for a method called isCarnivorous(), and then checks that it equals true or, in our case, false, because we're using should not.

Let's write one more example before we see this. How about: it_should_allow_to_check_if_dinosaur_is_carnivorous(). Inside, use $ this->beConstructedWith(). Remember, the Dinosaur class's constructor allows us to control the $ isCarnivorous value via the second argument. So, we'll pass 'Velociraptor' and true. Then we can say $ this->shouldBeCarnivorous(). That's the same thing as above, but without the 'not'.