Let's check it out! ./vendor/bin/phpspec run Woh! The error from our two new examples is: method [array:2] not found
Um... cool... so what the heck does that mean? It's not a great error. But, the question at the bottom tells us what is really going on:

  Do you want me to create Dinosaur::isCarnivorous() for you?

This is really cool! When we say shouldBeCarnivorous(), the ObjectStateMatcher says:

Hey! You're describing that your object 'should be carnivorous'. To figure that out, your class probably needs an isCarnivorous() method! So, let's create that thing!

Choose yes to generate it. phpspec re-runs and does fail - something about the ObjectStateMatcher expecting a boolean, but null given. That's because our new method is empty!