ObjectStateMatcher with 'shouldHave'
Now that our tests are green, we get to think about any refactoring we might want to do. Here's one piece I don't love: this logic for comparing whether or not the diet of two Dinosaurs is the same - it's just not super clear.

So, it might be nice to have a method on the Dinosaur class called hasSameDietAs(): we pass it a Dinosaur and it returns a boolean.

So... cool! Let's add an example for this: it_should_allow_to_check_if_two_dinosaurs_have_same_diet. And, check out the language $ this->shouldHaveSameDietAs(new Dinosaur()).

Two important things here. First, the  object will be a Dinosaur object that's created with no constructor args - so it will be a veggiesaurus. And so, it should have the same diet as a new Dinosaur(), which will also be non-carnivorous.

Second, see this language - shouldHaveSameDietAs()? That language will also be handled by the ObjectStateMatcher. Yep, when you say shouldBeSomething, it looks for an isSomething() method. And if you say shouldHaveSomething(), it looks for a hasSomething() method.

One of the problems I originally had with the ObjectStateMatcher was that I was thinking about it backwards. I was thinking:

  Hey! I want to have a method called isCarnivorous().

Then, I would try to figure out the correct matcher method to use - like shouldBeCarnivorous() - so that it would look for this method. But really, we need to think about it the other direction: I shouldn't care what the method name will be called in Dinosaur. Nope, I can ignore that and focus on using natural language in my example: $ this->shouldBeCarnivorous() and down here $ this->shouldHaveSameDietAs(). Use natural language, and then... don't even think about the method name! Just run phpspec - it'll tell you:
./vendor/bin/phpspec run

There it is! hasSameDietAs(). Generate that, then go find it. This method will return a bool, the argument will be a Dinosaur object and we can return $ dinosaur->isCarnivorous() === $ this->isCarnivorous().

Let's try it!
./vendor/bin/phpspec run
We are green!