In the Exception/ directory that we copied a few minutes ago, there's another exception class: the DinosaursAreRunningRampantException.
 8 lines  src/Exception/DinosaursAreRunningRampantException.php
final class DinosaursAreRunningRampantException extends \Exception
{
}
Here's the problem we're facing: we have these enclosures, but... they don't have any security - no electric fences, no guard towers, nothing! We need to add that capability to enclosures and throw this new exception if we try to add a Dinosaur to an Enclosure that has no active security. Because... honestly... we're having a real problem where people add dinosaurs to an enclosure and then just leave the door wide open.

In EnclosureSpec, let's create a new example to describe this: it_should_not_allow_to_add_dinosaurs_to_unsecure_enclosures(). I want you to temporarily ignore all the other examples that we've been working on so far, because this example is going to temporarily break... all of them.