First... how does 'security' for our enclosures need to be designed? Is it a boolean property on Enclosure so we can just turn security on or off? Something more complex? Actually, right now, it doesn't matter!
Check it out: in this example, I want to describe that if you simply create a new Enclosure, that is not enough: it's not secure. Describe that by saying $ this->shouldThrow(). And this time, instead of passing the class name of the exception that should be thrown, I'll say new DinosaursAreRunningRampantException() and pass this a message: Are you craaazy?.

Why am I doing this differently than before? It's really up to you: you can pass the class name to shouldThrow() if you only need to make sure the exception is an instance of that class or if you want to make sure that the message is also correct, you can create the exception object with the message you expect.

Next, the exception should be thrown ->duringAddDinosaur(new Dinosaur()) with Velociraptor and true.

Oh, and this language is also a bit different than before. Earlier, we used during() and passed addDinosaur as an argument to that method. That's fine, but you can also use this more magical way: duringAddDinosaur(). It's a bit more natural because you can then pass each argument one-by-one, instead of putting them in an array. They do the same thing - so it's up to you.

The point is: we now have a test that describes that you can't just create an Enclosure and start putting dinosaurs into it. Somehow, security needs to be activated... whatever that means.

Let's move over and run phpspec:
./vendor/bin/phpspec run
Awesome! That does fail because it is still possible to add dinosaurs to enclosures without activating security.