Designing the Security
Ok... so how do we need enclosure security to work? The right answer depends on your dinosaur park. But the process is universal: think about what requirements you have. Is security just something you turn on or off - like with some activateSecurity() method on Enclosure? Or, is it more complex? Based on talking to our security experts, I've determined that we need the ability to add different types of security to different Enclosures - electric fences around some, guard towers around others and maybe just a sign that says 'Please stay inside' if we get really busy. Oh, and each Enclosure can have 0 or many pieces of security.

Back in the tutorial/ directory, check out the Entity/ directory. See that Security class? Copy that and put into our src/Entity/ folder. Not spec/Entity, I'm totally messing this up right now... and will pay for it later.

And, yes, yes, we're going to cheat a bit: we're going to skip the spec process for the Security class and start with something I've already created.

Each Security has a name - like 'electric fence' or 'guard tower' and a boolean for whether it's active or not. And we pass in the Enclosure that this Security will be attached to. For the methods - just one: getIsActive().

To get the example in EnclosureSpec to pass, somehow, we need a way to attach Security objects to our Enclosure class. And then, when we add a dinosaur, we can check to make sure the Enclosure has at least one active Security.