To hold the securities, let's create a $ securities property and set it to an empty array. This will be an array of Security objects, so let's document that.
Now, this is interesting. If we're adding a securities property, shouldn't we describe this more directly with some examples that show... I don't know... some addSecurity() or getSecurities() methods? Well... maybe? Maybe because... we might not need these methods! Right now, what we do know is that, if there are no active securities, an exception should be thrown. And of course, we will need to update some of our examples from earlier once we get this working so that they also have some active security.

Anyways, down in addDinosaur(), let's call another new method if (! $ this->isSecurityActive()) we will throw a new DinosaursAreRunningRampantException() and pass it the same message that we described in our example - because we're testing for this exact string.

In reality, this is a bit silly. In real life, I probably wouldn't care enough to test for that exact message - the class is enough.

To add the missing method, I'll put my cursor on the method name, hit Alt + Enter and click 'Add Method'. Cool! This will return a bool and inside, we can loop over the $ securities with $ this->securities as $ security. If at least one Security object attached to this enclosure is active, then our enclosure is secure. So if ($ security->getIsActive()), then return true. And if none of them are active, return false.

Okay, that should work! Move back to the terminal. Oh, see this 41? That means the example lives on line 41 of the spec class. Re-run phpspec:
./vendor/bin/phpspec run

It works! Sort of. Notice, line 41 is gone - that example is passing! By the way, instead of running all your spec classes, you can run just one by passing the filename to the command:
./vendor/bin/phpspec run spec/Entity/EnclosureSpec.php
Or, you can run just one example by adding colon then the line number. The example we're working on should be line 38 - yep! There it is. Try it:
./vendor/bin/phpspec run spec/Entity/EnclosureSpec.php:38
Cool! 1 passed. But if we run all of them, we have a few failures.

We made a few changes to our app that broke our existing examples. Next, let's think about the correct way to handle this and add a few more nice features to our Enclosure... including testing exceptions that happen during object construction.