The new example we just added is passing... but we totally broke a bunch of our original examples! Lame! For example, it should be able to add dinosaurs is getting the DinosaursAreRunningRampantException and that makes sense. Find it_should_be_able_to_add_dinosaurs. It's testing to make sure that if we add two dinosaurs, then we should have... 2 dinosaurs.
Existing Tests Broke... now what?

So, hmm: we made a change to our app and then an existing example started to fail. That's... the beauty of tests! Now we can take in all this information about which tests are failing and why they're failing and determine the best way forward. For example, we may have just accidentally introduced a bug and we want to fix that bug. Or we may discover that the tests that are failing are no longer relevant and should be removed. Or, most likely you're in a situation like this one: where you simply need to update existing tests for some new change.

Describing adding Security
For this example, the enclosure needs some security before we starting adding dinosaurs. To do that, well... we need some way to add Security to an Enclosure! And, ya know what? We should probably make it pretty easy to add security - maybe via a new constructor arg.

Let me show you want I mean - in an example! Instead of creating a totally new example function for this, I'm going to need to use this new functionality in the example that's current failing. Here's the idea: to add some basic security, say $ this->beConstructedWith(true). Yep, I want there to be a constructor arg that easily allows you to activate some type of security.
Let's also add this to the other example that's failing - it's around line 32. Paste! And for the newest example we've been working on, I'll instantiate with false.