Ok, let's try the tests: ./vendor/bin/phpspec run
Nice! It asks us to generate the __construct method - yes please! A whole bunch of examples are failing - but, pfff - that's probably fine. Find Enclosure. Perfect!

Change the argument to bool $ withBasicSecurity and... I don't need to, but let's give this a default value: if you pass nothing, there is no security. Next, if (), let's... add some security! I'll call a new method we haven't created yet: $ this->addSecurity() with new Security() passing that fence... or I guess 'fency', whatever that is... true to make it active and  because it will be attached to this Enclosure.

For the addSecurity() method, because we're not using this method anywhere outside of this class, we should technically create it as private. But, I already know that I will need to use it outside this class, so let's make it public: public function addSecurity(Security $ security). Inside, $ this->securities[] = $ security.

Phew! Okay, find your terminal and let's try this!
./vendor/bin/phpspec run
Hmm, not passing yet:

Class App\Entity\Security not found.

Let's see, what did I mess up? It's not clear where that error is. To get more info, use the --verbose flag:
./vendor/bin/phpspec run --verbose
And... ah! There it is: Enclosure at line 19. Find that and... ah again! Our spec/ and src/ directories look so much alike that I copied the Security class into the wrong spot! Move that into src/Entity - good job tests!

Run 'em again:
./vendor/bin/phpspec run
Now they pass.