Exception during Construction?
We're on a roll! Shall we add one more enhancement to Enclosure? It's now easy to create an Enclosure with basic security. But I also want the ability to pass some initial dinosaurs into the constructor. That's cool - but another programmer tried to do this last week and... oof, things got ugly. They did allow for initial dinosaurs to be added, but they forgot to check first to see if any security was active. Oof. Anyways, that programmer is... 'unavailable' now.

Let's not make the same mistake: create a new example function: it_should_fail_if_providing_initial_dinosaurs_without_security(). Start with $ this->beConstructedWith(false) and an array with one new Dinosaur().

Then, we just need to tell phpspec what method will cause the exception. So... wait! This is a bit different: the exception will be thrown during instantiation! Not when we call some other method.

How can we tell phpspec about that? It's almost the same: $ this->shouldThrow() with DinosaursAreRunningRampantException::class - I don't care about testing the exact message. Then, ->duringInstantiation().

That's it. Let's make sure things are failing!
./vendor/bin/phpspec run
Nice and broken: there was no exception thrown