It's finally time to talk about one of the most critical parts of unit testing: mocking. Oh, and it's kind of the most fun part too!
To Mock or Not to Mock

Check out EnclosureSpec: we already had at least one situation where we called a method and needed to pass another object as an argument - a Dinosaur in this case. When the object you're testing has a dependency on another object like this, you have two options. First, you can just pass the real object, and that's what we've been doing so far. This is a simple and excellent solution when the object you're passing is easy to instantiate and doesn't have any side effects - most commonly, objects that just hold data.

The second solution - mocking - is perfect for all the other situations: when the object you're using is a pain to instantiate, its behavior is complex or its methods do things - like it makes database queries. In those cases, we do not want to use the real object: we want to mock... or create a test double... or a dummy. These are all terms that sorta describe the same thing - we'll discuss as we go along.

Creating a Mock Object

Let's see this in action! Create a new example: function it_should_allow_to_check_if_two_dinosaurs_have_same_diet_using_stub(). Yea, we'll discuss that word 'stub' along the way.

Check this out: instead of creating a new Dinosaur object, add an argument to the example method with a Dinosaur type-hint. Let's var_dump($ dinosaur) and then see what happens when we run phpspec:
./vendor/bin/phpspec run
Oh... interesting! It's some sort of a Collaborator object. But what I really want you to see is that, inside of it, is something called an ObjectProphecy. Woh, cool name.

Technically speaking, phpspec doesn't have its own mocking system - it uses a totally independent library called prophecy. Well, the truth is that the phpspec team made and maintains both libraries - but prophecy is its own library, and can even be used in PHPUnit.

But the point is, this is not a real Dinosaur object, it's a 'fake' object that looks and smells like a Dinosaur object and one that we can completely control. And getting a mock object is easy! Just add an argument type-hinted with the class or interface you need to mock - phpspec & prophecy take care of the rest. I love that.